LayoutTests/imported/w3c/web-platform-tests/event-timing/crossiframe-expected.txt

Generate a 'click' event

FAIL Event Timing: entries should only be observable by its own frame. assert_equals: There should be a single first-input entry expected 1 but got 0
